Renamed the completion trigger handler to be more specific about its role now that there are multiple typedHandler implementations. Ensured that the client-side on-type formatting logic is ordered before the server-side on-type formatting logic.
